refactor knox.js for clarity
