+++ src/main/resources/static/index.html
@@ -106,9 +106,9 @@
     </section>
 
     <section id="version">
-      
+
     </section>
-    
+
     <footer>
       <script src="vendor/sweet/sweetalert.min.js"></script>
       <script src="vendor/jquery/jquery.min.js"></script>
@@ -117,7 +117,8 @@
       <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.3/jquery.easing.min.js"></script>
       <script src="vendor/magnific-popup/jquery.magnific-popup.min.js"></script>
       <script type="text/javascript" src="vendor/tooltipster/tooltipster.bundle.min.js"></script>
-      <script src="js/knox.js"></script>
+      <script type="module" src="js/knox.js"></script>
+      <script type="module" src="js/target.js"></script>
     </footer>
   </body>
 </html>
